refactor sam service into service + dao
